refactor(domain): move AudioTrack and SubtitleTrack to media context
Track value objects describe physical media file metadata and belong
in the media bounded context. Library re-exports for backward compat.
TrackSelector now imports directly from media/shared.
